Exposed GameObject::IsEnabled to Lua. ComponentLuaScript will only register for keyboard events if OnKeys() is defined in the script. VisualScriptNode_Disable_GameObject toggles the status for now.

+++ MyEngine/SourceEditor/Documents/EditorDocument.cpp
@@ -266,7 +266,7 @@ void EditorDocument::SetRelativePath(const char* relativePath)
         m_Filename = m_RelativePath;
     }
 
-    for( int i=strlen( m_RelativePath ); i>=0; i-- )
+    for( int i=(int)strlen( m_RelativePath ); i>=0; i-- )
     {
         if( m_RelativePath[i] == '\\' || m_RelativePath[i] == '/' )
         {
